requirements.txt
antlr4-python3-runtime==4.9.3
-e git+https://github.com/mit-ll-trusted-autonomy/cbfToolbox@aa2d5843683cc2191b20241f2c52a07b70054a67#egg=cbf_toolbox
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gurobipy==11.0.3
hydra-core==1.3.2
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.14.0
six==1.16.0
sympy==1.13.1
torch==2.4.0
triton==3.0.0
typing_extensions==4.12.2
